requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
ciscoaxl==0.158
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
duo-client==4.3.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
heroku==0.1.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.3
MarkupSafe==1.1.1
meraki==1.4.3
MouseInfo==0.1.3
multidict==5.1.0
netmiko==3.4.0
ntc-templates==2.0.0
numpy==1.20.2
opencv-python==4.5.1.48
packaging @ file:///private/tmp/sip--packaging-20210330-83403-xxi7dp/packaging-20.9
paramiko==2.7.2
Pillow==8.2.0
protobuf==3.17.1
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyJWT==1.7.1
PyMsgBox==1.0.9
PyNaCl==1.4.0
pynput==1.7.3
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-sip==12.8.1
PyRect==0.1.4
PyScreeze==0.1.27
pyserial==3.5
pytesseract==0.3.7
python-dateutil==1.5
python-dotenv==0.14.0
PyTweening==1.0.3
pytz==2021.1
PyYAML==5.3.1
QScintilla==2.12.1
requests==2.25.0
requests-toolbelt==0.9.1
rubicon-objc==0.4.0
scp==0.13.3
selenium==3.141.0
sip==6.0.3
six==1.12.0
smartsheet-python-sdk==2.105.1
soupsieve==2.2.1
TBB==0.1
tenacity==7.0.0
tesseract==0.1.3
textfsm==1.1.0
thread-task==0.9.7
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.2
virtualenv==20.4.3
webexteamssdk==1.6
Werkzeug==1.0.1
whitenoise==5.2.0
withings-sync==3.1.3
yarl==1.6.3